+++ src/pxfquery/__init__.py
@@ -1,5 +1,5 @@
 from pxfquery._version import __version__
-from pxfquery.client import PxFQuery
+from pxfquery.presentation.client import PxFQuery
 
 __all__ = [
     "PxFQuery",