+++ src/pxfquery/api.py
@@ -1,3 +1,0 @@
-from pxfquery.query import parse, query
-
-__all__ = ["query", "parse"]